modification du typage de l'alt pour etre plus cohérent, ajout de la doc du simulator, ajout du calcul de la vitesse verticale avec un liftCoef constant pour l'instant

+++ src/core/Simulator/Simulator.h
@@ -21,7 +21,7 @@ private:
     void simLoop();
     void computeGroundSpeed();
     void computeVerticalSpeed();
-    double computeAirDensity(int altitude);
+    double computeAirDensity(double altitude);
 
 public:
     // Constructeur
